requirements.txt
asgiref==3.6.0
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-bootstrap-datepicker-plus==5.0.3
django-ckeditor==6.5.1
django-environ==0.9.0
django-import-export==3.1.0
django-js-asset==2.0.0
django-modeltranslation==0.18.9
django-online-users==0.3
djangorestframework==3.14.0
docx2pdf==0.1.8
et-xmlfile==1.1.0
filetype==1.2.0
import-export==0.3.1
lxml==4.9.2
MarkupPy==1.14
numpy==1.24.2
odfpy==1.4.1
openpyxl==3.1.1
Pillow==9.4.0
psycopg2==2.9.5
pydantic==1.10.5
pypdf2==3.0.1
python-docx==0.8.11
pytz==2023.3
PyYAML==6.0
reportlab==3.6.12
sqlparse==0.4.3
tablib==3.3.0
tqdm==4.64.1
typing-extensions==4.5.0
tzdata==2022.7
xlrd==2.0.1
xlwt==1.3.0
